I originally wrote the persistence / Portfolio.java @Entity as a JavaFx model with Properties.  It didn't work as Properties aren't serializable, so I reverted and planned to use a DTO to load between the model and database.  I realised something and moved the @Id from the 'long id' to the 'getId()' and then suddenly the persistence not happening on the property but the getters that return a string.  So this works now and don't need to have separate model and persistencee entities.  However I always planned to have some separation so that I can quite easily convert to using a WebApp with web services to deliver the data instead of it coming from a local Db.
What I think I'll do here is retain the Model and Persistencee, and put a backend implementation specific DTO (based on an Interface which I already have - its PortfolioLoader) which for the current arrangement with a local DB will have the two properties (ie. one for each at each end) mirror the other like properties do.  I'm not sure if you can do this so will do some investigation.

Anyway, this commit is having the new persistence/Portfolio.java @Entity that is composed of properties.